cheri: Manually apply fixes for FixPtrToULongCast warnings
Manually add fixes for the FixPtrToULongCast warnings where
clang-tidy isn't able to fix them automatically due to macro
expansions.

Explicitly ignore warnings in fs/btrfs/ because this file system
plays tricks with pointers that need a different fix.

Ignore some warnings that need a different fix.

+++ arch/riscv/kernel/traps_misaligned.c
@@ -143,7 +143,7 @@
 	(SHIFT_RIGHT((insn), (pos) - LOG_REGBYTES) & REG_MASK)
 
 #define REG_PTR(insn, pos, regs)	\
-	(ulong *)((ulong)(regs) + REG_OFFSET(insn, pos))
+	(ulong *)((uintptr_t)(regs) + REG_OFFSET(insn, pos))
 
 #define GET_RS1(insn, regs)		(*REG_PTR(insn, SH_RS1, regs))
 #define GET_RS2(insn, regs)		(*REG_PTR(insn, SH_RS2, regs))
